+++ get_x_navigation_generator/lib/src/resolvers/importable_type_resolver.dart
@@ -14,8 +14,7 @@ abstract class ImportableTypeResolver {
     bool forceNullable = false,
   });
 
-  ImportableType resolveFunctionType(FunctionType function,
-      [ExecutableElement? executableElement]);
+  ImportableType resolveFunctionType(FunctionType function, [ExecutableElement? executableElement]);
 
   static String? relative(String? path, Uri? to) {
     if (path == null || to == null) {
@@ -23,16 +22,11 @@ abstract class ImportableTypeResolver {
     }
     var fileUri = Uri.parse(path);
     var libName = to.pathSegments.first;
-    if ((to.scheme == 'package' &&
-            fileUri.scheme == 'package' &&
-            fileUri.pathSegments.first == libName) ||
-        (to.scheme == 'asset' && fileUri.scheme != 'package')) {
+    if ((to.scheme == 'package' && fileUri.scheme == 'package' && fileUri.pathSegments.first == libName) || (to.scheme == 'asset' && fileUri.scheme != 'package')) {
       if (fileUri.path == to.path) {
         return fileUri.pathSegments.last;
       } else {
-        return p.posix
-            .relative(fileUri.path, from: to.path)
-            .replaceFirst('../', '');
+        return p.posix.relative(fileUri.path, from: to.path).replaceFirst('../', '');
       }
     } else {
       return path;
@@ -64,8 +58,7 @@ class ImportableTypeResolverImpl extends ImportableTypeResolver {
     }
 
     for (var lib in libs) {
-      if (!_isCoreDartType(lib) &&
-          lib.exportNamespace.definedNames.values.contains(element)) {
+      if (!_isCoreDartType(lib) && lib.exportNamespace.definedNames.values.contains(element)) {
         return lib.identifier;
       }
     }
@@ -77,10 +70,8 @@ class ImportableTypeResolverImpl extends ImportableTypeResolver {
   }
 
   @override
-  ImportableType resolveFunctionType(FunctionType function,
-      [ExecutableElement? executableElement]) {
-    final functionElement =
-        executableElement ?? function.element2 ?? function.alias?.element;
+  ImportableType resolveFunctionType(FunctionType function, [ExecutableElement? executableElement]) {
+    final functionElement = executableElement ?? function.element ?? function.alias?.element;
     if (functionElement == null) {
       throw 'Can not resolve function type \nTry using an alias e.g typedef MyFunction = ${function.getDisplayString(withNullability: false)};';
     }
@@ -88,7 +79,7 @@ class ImportableTypeResolverImpl extends ImportableTypeResolver {
     var functionName = displayName;
 
     Element elementToImport = functionElement;
-    var enclosingElement = functionElement.enclosingElement3;
+    var enclosingElement = functionElement.enclosingElement;
 
     if (enclosingElement != null && enclosingElement is ClassElement) {
       functionName = '${enclosingElement.displayName}.$displayName';
@@ -106,13 +97,12 @@ class ImportableTypeResolverImpl extends ImportableTypeResolver {
     final importableTypes = <ImportableType>[];
     if (typeToCheck is ParameterizedType) {
       for (DartType type in typeToCheck.typeArguments) {
-        if (type.element2 is TypeParameterElement) {
+        if (type.element is TypeParameterElement) {
           importableTypes.add(const ImportableType(className: 'dynamic'));
         } else {
           importableTypes.add(ImportableType(
-            className: type.element2?.name?.replaceAll('?', '') ??
-                type.getDisplayString(withNullability: false),
-            import: resolveImport(type.element2),
+            className: type.element?.name?.replaceAll('?', '') ?? type.getDisplayString(withNullability: false),
+            import: resolveImport(type.element),
             isNullable: type.nullabilitySuffix == NullabilitySuffix.question,
             typeArguments: _resolveTypeArguments(type),
           ));
@@ -130,12 +120,10 @@ class ImportableTypeResolverImpl extends ImportableTypeResolver {
     bool forceNullable = false,
   }) {
     return ImportableType(
-      className:
-          type.element2?.name ?? type.getDisplayString(withNullability: false),
+      className: type.element?.name ?? type.getDisplayString(withNullability: false),
       name: name,
-      isNullable:
-          forceNullable || type.nullabilitySuffix == NullabilitySuffix.question,
-      import: resolveImport(type.element2),
+      isNullable: forceNullable || type.nullabilitySuffix == NullabilitySuffix.question,
+      import: resolveImport(type.element),
       isRequired: isRequired,
       typeArguments: _resolveTypeArguments(type),
     );